+++ R/distionary-package.R
@@ -95,5 +95,7 @@
 "_PACKAGE"
 
 ## usethis namespace: start
+#' @importFrom Rcpp sourceCpp
+#' @useDynLib distionary, .registration = TRUE
 ## usethis namespace: end
 NULL